requirements.txt
asgiref==3.6.0
asttokens==2.4.1
attrs==23.2.0
autobahn==19.5.1
autograd==1.6.2
autograd-gamma==0.5.0
Automat==22.10.0
bleach==3.1.5
Brotli==1.1.0
captum==0.7.0
certifi==2024.8.30
cffi==1.16.0
channels==2.3.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
daphne==2.4.1
debugpy==1.8.2
decorator==5.1.1
Django==2.2.28
djangorestframework==3.9.4
docker-pycreds==0.4.0
ecos==2.0.14
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.13.1
fonttools==4.53.1
formulaic==1.0.2
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.2.1
h5py==3.11.0
hyperlink==21.0.0
idna==3.8
incremental==22.10.0
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
interface-meta==1.3.0
ipykernel==6.29.5
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jsonfield2==3.0.3
jsonschema==3.0.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lifelines==0.29.0
llvmlite==0.43.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mkl==2024.2.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
munch==4.0.0
mypy_extensions==0.4.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
osqp==0.6.7.post1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==1.26.0
pyrsistent==0.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
qdldl==0.1.7.post4
requests==2.32.3
roman==3.1
scikit-learn==1.5.1
scikit-survival==0.23.0
scipy==1.13
seaborn==0.13.2
sentry-sdk==2.12.0
service-identity==24.1.0
setproctitle==1.3.3
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.12
tables==3.10.0
tbb==2021.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
topk @ git+https://github.com/oval-group/smooth-topk.git@12c1645f187e2fa0c05f47bf1fe48864d4bd2707
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==3.0.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==2.2.3
wandb==0.17.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==8.1
wrapt==1.16.0
zope.interface==6.4.post2
